feat: adicionando ultimos ajustes
