+++ src/components/Button/Button.jsx
@@ -1,0 +1,5 @@
+function Button () {
+    return (
+        <></>
+    )
+}
