+++ src/components/Header/index.css
@@ -1,0 +1,3 @@
+.title h1{
+font-size: 25px;
+}
