+++ src/pages/Home/index.css
@@ -1,0 +1,24 @@
+#root {
+    max-width: 1280px;
+    margin: 0 auto;
+    padding: 2rem;
+    text-align: center;
+  }
+  .card {
+    padding: 2em;
+  }
+  
+  .read-the-docs {
+    color: #888;
+  }
+  
+  .group{
+  display: flex;
+  gap: 1rem;
+  margin-top: 1rem;
+    
+  }
+   .timer{
+    font-size: 2rem;
+    color: #646cff;
+   }
